RAG_mlflow/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.25.2
mlflow==2.7.2
